test: add integration test, unit test

+++ src/main/java/com/Man10h/book_store/service/impl/UserServiceImpl.java
@@ -1,8 +1,10 @@
 package com.Man10h.book_store.service.impl;
 
+import com.Man10h.book_store.exception.business.AccountDisabledException;
 import com.Man10h.book_store.exception.business.UserNotFoundException;
 import com.Man10h.book_store.exception.ErrorException;
 import com.Man10h.book_store.model.dto.ChatMessage;
+import com.Man10h.book_store.model.entity.UserEntity;
 import com.Man10h.book_store.model.response.UserResponse;
 import com.Man10h.book_store.repository.UserRepository;
 import com.Man10h.book_store.service.UserService;
@@ -14,6 +16,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -30,12 +34,16 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @CacheEvict(value = "users", allEntries = true)
     public void updateUserRole(Long id) {
+        Optional<UserEntity> optional = userRepository.findById(id);
+        if(optional.isEmpty()){
+            throw new UserNotFoundException("User not found");
+        }
+        if(!optional.get().getEnabled()){
+            throw new AccountDisabledException("Account not enabled");
+        }
         try{
             userRepository.updateUserRole(id, 2L);
-        }catch (UserNotFoundException e){
-            throw new UserNotFoundException(e.getMessage());
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             throw new ErrorException(e.getMessage());
         }
     }
@@ -43,10 +51,12 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @CacheEvict(value = "users", allEntries = true)
     public void deleteUser(Long id) {
+        Optional<UserEntity> optional = userRepository.findById(id);
+        if(optional.isEmpty()){
+            throw new UserNotFoundException("User not found");
+        }
         try{
             userRepository.deleteById(id);
-        }catch (UserNotFoundException e){
-            throw new UserNotFoundException(e.getMessage());
         }
         catch (Exception e) {
             throw new ErrorException(e.getMessage());
